# Conflicts:
#	Client/src/services/EventService.js
#	Client/src/views/main.js

+++ Server/src/routes/organiserRoutes.js
@@ -16,7 +16,7 @@ let router = express.Router();
 // TODO add auth to all this shit
 
 // Middleware for organiser activities BRUK DENNE FOR USER OGSÅ
-/*app.use("/organiser", (req, res, next) => {
+/*app.use("", (req, res, next) => {
   var token = req.headers["x-access-token"];
   let decoded = td.decode(token);
   if (decoded.status == 200) {
@@ -53,13 +53,23 @@ router.get('/event/:id', (req: express$Request, res: express$Response) => {
 
 // Create new event (and connect it to the organiser)
 router.post('/event', (req: { body: Object }, res: express$Response) => {
-  dao.postEvent(req.body, (status, data) => {
-    res.status(status);
-    let d = data;
-    dao.postEventOrganiser(data.insertId, (status, data) => {
-      res.status(status);
-      res.send(d);
-    });
+  td.decode(req.headers['x-access-token'], (err, decoded) => {
+    if (err) {
+      res.status(401);
+      res.send(err);
+    } else {
+      dao.postEvent(req.body, (status, data) => {
+        if (status == 200) {
+          dao.postEventOrganiser(data.insertId, decoded.username, (status, data) => {
+            res.status(status);
+            res.send(data);
+          });
+        } else {
+          res.status(status);
+          res.send(data);
+        }
+      });
+    }
   });
 });
 
@@ -76,37 +86,53 @@ router.put('/event', (req: { body: Object }, res: express$Response) => {
 
 // Delete single event
 router.delete('/event/:id', (req: express$Request, res: express$Response) => {
-  dao.deleteEventOrganisers(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
-  });
-  dao.deleteEventVolunteers(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
-  });
-  dao.deleteEventArtists(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
-  });
-  dao.deleteEventFiles(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
-  });
-  dao.deleteEventTickets(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
-  });
-  dao.deleteEventRiders(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
-  });
-  dao.deleteEventSchedule(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
-  });
-  dao.deleteEvent(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
+  td.decode(req.headers['x-access-token'], (err, decoded) => {
+    if (err) {
+      res.status(401);
+      res.send(err);
+    }
+    let organiserEmail = decoded.username;
+    // check if organiser the organiser has an event with the provided id
+    dao.organiserOwnsEvent(req.params.id, organiserEmail, (status, data) => {
+      if (data.length == 0) {
+        res.status(404);
+        res.send({ error: 'Arragementet eksiterer ikke' });
+      } else {
+        var completedDeletions = 0;
+        function checkIfDone() {
+          if (completedDeletions > 6) {
+            res.status(200);
+            res.send('Event deleted');
+          } else {
+            completedDeletions++;
+          }
+        }
+        dao.deleteEventOrganisers(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+        dao.deleteEventVolunteers(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+        dao.deleteEventArtists(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+        dao.deleteEventFiles(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+        dao.deleteEventTickets(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+        dao.deleteEventRiders(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+        dao.deleteEventSchedule(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+        dao.deleteEvent(req.params.id, (status, data) => {
+          checkIfDone();
+        });
+      }
+    });
   });
 });
 
@@ -126,11 +152,18 @@ router.post('/artist/:id', (req: { body: Object }, res: express$Response) => {
   });
 });
 
-// Get a group of volunteers from an organiser.
-router.get('/group/:id', (req: express$Request, res: express$Response) => {
-  dao.getGroup(req.params.id, (status, data) => {
-    res.status(status);
-    res.send(data);
+// Get all types of volunteers from an organiser.
+router.get('/group', (req: express$Request, res: express$Response) => {
+  td.decode(req.headers['x-access-token'], (err, decoded) => {
+    if (err) {
+      res.status(401);
+      res.send(err);
+    } else {
+      dao.getGroup(decoded.username, (status, data) => {
+        res.status(status);
+        res.send(data);
+      });
+    }
   });
 });
 
@@ -228,6 +261,14 @@ router.get('/event/:id/volunteer', (req: express$Request, res: express$Response)
 //Get all artists who are part of an event
 router.get('/event/:id/artist', (req: express$Request, res: express$Response) => {
   dao.getArtistsByEvent(req.params.id, (status, data) => {
+    res.status(status);
+    res.send(data);
+  });
+});
+
+// Edit a ticket type
+router.put('/tickets', (req: { body: Object }, res: express$Response) => {
+  dao.editTicketType(req.body, (status, data) => {
     res.status(status);
     res.send(data);
   });
